Move StartedThreadCpuTimer to new package for common-cpp migration
