+++ src/main/cc/wfa/measurement/common/time/started_thread_cpu_timer.h
@@ -12,15 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef WFA_MEASUREMENT_COMMON_CRYPTO_STARTED_THREAD_CPU_TIMER_H_
-#define WFA_MEASUREMENT_COMMON_CRYPTO_STARTED_THREAD_CPU_TIMER_H_
+#ifndef WFA_MEASUREMENT_COMMON_TIME_STARTED_THREAD_CPU_TIMER_H_
+#define WFA_MEASUREMENT_COMMON_TIME_STARTED_THREAD_CPU_TIMER_H_
 
-#include <time.h>
+#include <ctime>
 
 #include "absl/time/time.h"
 #include "glog/logging.h"
 
-namespace wfa::measurement::common::crypto {
+namespace wfa {
 
 namespace {
 
@@ -50,6 +50,6 @@ class StartedThreadCpuTimer {
   const absl::Duration start_;
 };
 
-}  // namespace wfa::measurement::common::crypto
+}  // namespace wfa
 
-#endif  // WFA_MEASUREMENT_COMMON_CRYPTO_STARTED_THREAD_CPU_TIMER_H_
+#endif  // WFA_MEASUREMENT_COMMON_TIME_STARTED_THREAD_CPU_TIMER_H_
